fix i210_replay and get_table_disk

+++ flow/data_pipeline/leaderboard_utils.py
@@ -63,10 +63,12 @@ def get_table_disk(table_name="fact_vehicle_trace", bucket="circles.data.pipelin
     keys = [e["Key"] for e in response["Contents"] if e["Key"].find(table_name) == 0 and e["Key"][-4:] == ".csv"]
     names = [key_to_name(k) for k in keys]
     existing_results = os.listdir("./result/{}".format(table_name))
+    updated = False
     for index in range(len(keys)):
         if names[index] not in existing_results:
+            updated = True
             s3.download_file(bucket, keys[index], "./result/{}/{}".format(table_name, names[index]))
-    if table_name == "leaderboard_chart_agg":
+    if table_name == "leaderboard_chart_agg" and updated:
         for p in existing_results:
             os.remove("./result/{}/{}".format(table_name, p))
